renamed kuzzleRoom.js -> Room.js
